refactor(gmail): test code and gmail code
