+++ .env.testing.example
@@ -36,6 +36,8 @@ REDIS_HOST=127.0.0.1
 REDIS_PASSWORD=null
 REDIS_PORT=6379
 
+GMAIL_FROM_ADDRESS="jrnjdeveloper@gmail.com"
+
 MAIL_MAILER=smtp
 MAIL_HOST=mailpit
 MAIL_PORT=1025
